test-requirements.txt
flake8==2.5.1
coverage==4.0.3
pytest==2.8.5
pytest-cov==2.2.0
pytest-flask==0.10.0
requests==2.9.1
